Add Parser and SortCommand to support sorting
Lack of Parser and SortCommand files to support sorting

Added a SortCommandParser class to parse parameters
passed for sort command. Sort command utilizes prefixes
f/ and o/ for field and order respectively.

Added a SortCommand class. SortCommand object contains
field and object attributes to denote the specific
field and object to be implemented.

Implementation will be done in the next commit.